Connect YesNo page (listening comprehension page)
